Moved implementations from headers to cpp
