+++ CMakeLists.txt
@@ -3,6 +3,8 @@ cmake_minimum_required(VERSION 3.16)
 project(mmpr VERSION 0.1.0 LANGUAGES CXX)
 
 set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
+set(CMAKE_CXX_EXTENSIONS OFF)
 
 if(CMAKE_SOURCE_DIR STREQUAL CMAKE_CURRENT_SOURCE_DIR)
     # in standalone set default settings to ON
@@ -14,32 +16,28 @@ else()
     option(MMPR_BUILD_EXAMPLES "Build examples" OFF)
 endif()
 
+set(ZSTD_BUILD_STATIC ON)
+set(ZSTD_BUILD_SHARED OFF)
+set(ZSTD_LEGACY_SUPPORT OFF)
+set(ZSTD_BUILD_PROGRAMS OFF)
+set(ZSTD_BUILD_TESTS OFF)
+add_subdirectory("${CMAKE_CURRENT_SOURCE_DIR}/libs/zstd/build/cmake")
+
 # mmpr library target
-file(GLOB_RECURSE MMPR_SRC_FILES src/*.cpp)
+file(GLOB_RECURSE MMPR_SRC_FILES CONFIGURE_DEPENDS src/*.cpp)
 add_library(mmpr ${MMPR_SRC_FILES})
 add_library(mmpr::mmpr ALIAS mmpr)
 
+target_compile_features(mmpr PUBLIC cxx_std_17)
 target_include_directories(mmpr
     PUBLIC
         $<INSTALL_INTERFACE:include>
-        $<INSTALL_INTERFACE:include/filesystem>
-        $<INSTALL_INTERFACE:include/filesystem/writing>
-        $<INSTALL_INTERFACE:include/filesystem/reading>
-        $<INSTALL_INTERFACE:include/pcap>
-        $<INSTALL_INTERFACE:include/pcapng>
-        $<INSTALL_INTERFACE:include/modified_pcap>
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include/filesystem>
-        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include/filesystem/writing>
-        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include/filesystem/reading>
-        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include/pcap>
-        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include/pcapng>
-        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include/modified_pcap>
     PRIVATE
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/libs/zstd/lib>
 )
+target_link_libraries(mmpr PRIVATE libzstd_static)
 
-target_compile_features(mmpr PUBLIC cxx_std_17)
 if(CMAKE_CXX_COMPILER_ID STREQUAL "MSVC")
     # Visual Studio
     target_compile_options(mmpr PRIVATE /Wall)
@@ -53,14 +51,6 @@ if((CMAKE_SOURCE_DIR STREQUAL CMAKE_CURRENT_SOURCE_DIR) AND (CMAKE_BUILD_TYPE ST
     # If compiling as stand-alone project in debug mode set debug flag
     target_compile_options(mmpr PRIVATE -DDEBUG)
 endif()
-
-set(ZSTD_BUILD_STATIC ON)
-set(ZSTD_BUILD_SHARED OFF)
-set(ZSTD_LEGACY_SUPPORT OFF)
-set(ZSTD_BUILD_PROGRAMS OFF)
-set(ZSTD_BUILD_TESTS OFF)
-add_subdirectory("${CMAKE_CURRENT_SOURCE_DIR}/libs/zstd/build/cmake")
-target_link_libraries(mmpr PRIVATE libzstd_static)
 
 if(WIN32)
     # Windows requires winsock32.h for ntohs and ntohl etc.
