+++ include/mmpr/filesystem/writing/FileWriter.hpp
@@ -9,7 +9,7 @@ namespace mmpr {
 
 class FileWriter {
 public:
-    FileWriter(const std::string& filepath) : mFilepath(filepath) {}
+    FileWriter(const std::string& filepath);
 
     virtual void write(const uint8_t* data, size_t size) = 0;
