+++ src/filesystem/writing/FileWriter.cpp
@@ -1,1 +1,7 @@
 #include "mmpr/filesystem/writing/FileWriter.hpp"
+
+namespace mmpr {
+
+FileWriter::FileWriter(const std::string& filepath) : mFilepath(filepath) {}
+
+} // namespace mmpr
