+++ src/modified_pcap/ModifiedPcapReader.cpp
@@ -1,1 +1,53 @@
 #include "mmpr/modified_pcap/ModifiedPcapReader.hpp"
+
+namespace mmpr {
+
+template <typename TReader>
+ModifiedPcapReader<TReader>::ModifiedPcapReader(const std::string& filepath)
+    : mReader(filepath) {
+    modified_pcap::FileHeader fileHeader{};
+    ModifiedPcapParser::readFileHeader(mReader.data(), fileHeader);
+    mReader.mOffset += 24;
+}
+
+template <typename TReader>
+ModifiedPcapReader<TReader>::ModifiedPcapReader(TReader&& reader)
+    : mReader(std::forward<TReader>(reader)) {
+    modified_pcap::FileHeader fileHeader{};
+    ModifiedPcapParser::readFileHeader(mReader.data(), fileHeader);
+    mReader.mOffset += 24;
+}
+
+template <typename TReader>
+bool ModifiedPcapReader<TReader>::isExhausted() const {
+    return mReader.isExhausted();
+}
+
+template <typename TReader>
+bool ModifiedPcapReader<TReader>::readNextPacket(Packet& packet) {
+    if (isExhausted()) {
+        // nothing more to read
+        return false;
+    }
+
+    // make sure there are enough bytes to read
+    if (mReader.getSafeToReadSize() < 24) {
+        throw std::runtime_error(
+            "Expected to read at least one more raw packet record (24 bytes "
+            "at least), but there are only " +
+            std::to_string(mReader.getSafeToReadSize()) + " bytes left in the file");
+    }
+
+    modified_pcap::PacketRecord packetRecord{};
+    ModifiedPcapParser::readPacketRecord(&mReader.data()[mReader.mOffset], packetRecord);
+    packet.timestampSeconds = packetRecord.timestampSeconds;
+    packet.captureLength = packetRecord.captureLength;
+    packet.length = packetRecord.length;
+    packet.data = packetRecord.data;
+
+    mReader.mOffset += 24 + packetRecord.captureLength;
+
+    return true;
+}
+
+} // namespace mmpr
